authzmq/requirements.txt
pyzmq==14.3.0
